Refactor images and pages
